+++ exporter/exporterhelper/factory_test.go
@@ -16,10 +16,8 @@ package exporterhelper
 
 import (
 	"context"
-	"errors"
 	"testing"
 
-	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 	"go.uber.org/zap"
 
@@ -53,8 +51,6 @@ func TestNewFactory(t *testing.T) {
 		defaultConfig)
 	assert.EqualValues(t, typeStr, factory.Type())
 	assert.EqualValues(t, defaultCfg, factory.CreateDefaultConfig())
-	_, ok := factory.(component.ConfigUnmarshaler)
-	assert.False(t, ok)
 	_, err := factory.CreateTracesExporter(context.Background(), component.ExporterCreateParams{Logger: zap.NewNop()}, defaultCfg)
 	assert.Equal(t, componenterror.ErrDataTypeIsNotSupported, err)
 	_, err = factory.CreateMetricsExporter(context.Background(), component.ExporterCreateParams{Logger: zap.NewNop()}, defaultCfg)
@@ -69,14 +65,9 @@ func TestNewFactory_WithConstructors(t *testing.T) {
 		defaultConfig,
 		WithTraces(createTraceExporter),
 		WithMetrics(createMetricsExporter),
-		WithLogs(createLogsExporter),
-		WithCustomUnmarshaler(customUnmarshaler))
+		WithLogs(createLogsExporter))
 	assert.EqualValues(t, typeStr, factory.Type())
 	assert.EqualValues(t, defaultCfg, factory.CreateDefaultConfig())
-
-	fu, ok := factory.(component.ConfigUnmarshaler)
-	assert.True(t, ok)
-	assert.Equal(t, errors.New("my error"), fu.Unmarshal(nil, nil))
 
 	te, err := factory.CreateTracesExporter(context.Background(), component.ExporterCreateParams{Logger: zap.NewNop()}, defaultCfg)
 	assert.NoError(t, err)
@@ -105,8 +96,4 @@ func createMetricsExporter(context.Context, component.ExporterCreateParams, conf
 
 func createLogsExporter(context.Context, component.ExporterCreateParams, config.Exporter) (component.LogsExporter, error) {
 	return nopLogsExporter, nil
-}
-
-func customUnmarshaler(*viper.Viper, interface{}) error {
-	return errors.New("my error")
 }