+++ receiver/otlpreceiver/config.go
@@ -15,9 +15,20 @@
 package otlpreceiver
 
 import (
+	"fmt"
+
+	"github.com/spf13/cast"
+
 	"go.opentelemetry.io/collector/config"
 	"go.opentelemetry.io/collector/config/configgrpc"
 	"go.opentelemetry.io/collector/config/confighttp"
+)
+
+const (
+	// Protocol values.
+	protoGRPC          = "grpc"
+	protoHTTP          = "http"
+	protocolsFieldName = "protocols"
 )
 
 // Protocols is the configuration for the supported protocols.
@@ -35,8 +46,50 @@ type Config struct {
 }
 
 var _ config.Receiver = (*Config)(nil)
+var _ config.CustomUnmarshable = (*Config)(nil)
 
 // Validate checks the receiver configuration is valid
 func (cfg *Config) Validate() error {
+	return nil
+}
+
+// Unmarshal a config.Parser into the config struct.
+func (cfg *Config) Unmarshal(componentParser *config.Parser) error {
+	if componentParser == nil || len(componentParser.AllKeys()) == 0 {
+		return fmt.Errorf("empty config for OTLP receiver")
+	}
+	// first load the config normally
+	err := componentParser.UnmarshalExact(cfg)
+	if err != nil {
+		return err
+	}
+
+	// next manually search for protocols in viper, if a protocol is not present it means it is disable.
+	protocols := cast.ToStringMap(componentParser.Get(protocolsFieldName))
+
+	// UnmarshalExact will ignore empty entries like a protocol with no values, so if a typo happened
+	// in the protocol that is intended to be enabled will not be enabled. So check if the protocols
+	// include only known protocols.
+	knownProtocols := 0
+	if _, ok := protocols[protoGRPC]; !ok {
+		cfg.GRPC = nil
+	} else {
+		knownProtocols++
+	}
+
+	if _, ok := protocols[protoHTTP]; !ok {
+		cfg.HTTP = nil
+	} else {
+		knownProtocols++
+	}
+
+	if len(protocols) != knownProtocols {
+		return fmt.Errorf("unknown protocols in the OTLP receiver")
+	}
+
+	if cfg.GRPC == nil && cfg.HTTP == nil {
+		return fmt.Errorf("must specify at least one protocol when using the OTLP receiver")
+	}
+
 	return nil
 }