+++ receiver/receiverhelper/factory_test.go
@@ -16,10 +16,8 @@ package receiverhelper
 
 import (
 	"context"
-	"errors"
 	"testing"
 
-	"github.com/spf13/viper"
 	"github.com/stretchr/testify/assert"
 
 	"go.opentelemetry.io/collector/component"
@@ -40,8 +38,6 @@ func TestNewFactory(t *testing.T) {
 		defaultConfig)
 	assert.EqualValues(t, typeStr, factory.Type())
 	assert.EqualValues(t, defaultCfg, factory.CreateDefaultConfig())
-	_, ok := factory.(component.ConfigUnmarshaler)
-	assert.False(t, ok)
 	_, err := factory.CreateTracesReceiver(context.Background(), component.ReceiverCreateParams{}, factory.CreateDefaultConfig(), nil)
 	assert.Error(t, err)
 	_, err = factory.CreateMetricsReceiver(context.Background(), component.ReceiverCreateParams{}, factory.CreateDefaultConfig(), nil)
@@ -56,14 +52,9 @@ func TestNewFactory_WithConstructors(t *testing.T) {
 		defaultConfig,
 		WithTraces(createTraceReceiver),
 		WithMetrics(createMetricsReceiver),
-		WithLogs(createLogsReceiver),
-		WithCustomUnmarshaler(customUnmarshaler))
+		WithLogs(createLogsReceiver))
 	assert.EqualValues(t, typeStr, factory.Type())
 	assert.EqualValues(t, defaultCfg, factory.CreateDefaultConfig())
-
-	fu, ok := factory.(component.ConfigUnmarshaler)
-	assert.True(t, ok)
-	assert.Equal(t, errors.New("my error"), fu.Unmarshal(nil, nil))
 
 	_, err := factory.CreateTracesReceiver(context.Background(), component.ReceiverCreateParams{}, factory.CreateDefaultConfig(), nil)
 	assert.NoError(t, err)
@@ -89,8 +80,4 @@ func createMetricsReceiver(context.Context, component.ReceiverCreateParams, conf
 
 func createLogsReceiver(context.Context, component.ReceiverCreateParams, config.Receiver, consumer.Logs) (component.LogsReceiver, error) {
 	return nil, nil
-}
-
-func customUnmarshaler(*viper.Viper, interface{}) error {
-	return errors.New("my error")
 }