+++ service/parserprovider/default.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//       http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,23 +12,11 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package fileexporter
+package parserprovider
 
-import (
-	"go.opentelemetry.io/collector/config"
-)
-
-// Config defines configuration for file exporter.
-type Config struct {
-	*config.ExporterSettings `mapstructure:"-"`
-
-	// Path of the file to write to. Path is relative to current directory.
-	Path string `mapstructure:"path"`
-}
-
-var _ config.Exporter = (*Config)(nil)
-
-// Validate checks if the exporter configuration is valid
-func (cfg *Config) Validate() error {
-	return nil
+// Default is the default ParserProvider and it creates configuration from a file
+// defined by the --config command line flag and overwrites properties from --set
+// command line flag (if the flag is present).
+func Default() ParserProvider {
+	return NewSetFlag(NewFile())
 }