+++ service/parserprovider/inmemory.go
@@ -4,7 +4,7 @@
 // you may not use this file except in compliance with the License.
 // You may obtain a copy of the License at
 //
-//       http://www.apache.org/licenses/LICENSE-2.0
+//      http://www.apache.org/licenses/LICENSE-2.0
 //
 // Unless required by applicable law or agreed to in writing, software
 // distributed under the License is distributed on an "AS IS" BASIS,
@@ -12,23 +12,28 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-package fileexporter
+package parserprovider
 
 import (
+	"io"
+
 	"go.opentelemetry.io/collector/config"
 )
 
-// Config defines configuration for file exporter.
-type Config struct {
-	*config.ExporterSettings `mapstructure:"-"`
-
-	// Path of the file to write to. Path is relative to current directory.
-	Path string `mapstructure:"path"`
+type inMemoryProvider struct {
+	buf io.Reader
 }
 
-var _ config.Exporter = (*Config)(nil)
+// NewInMemory returns a new ParserProvider that reads the configuration from the provided buffer as YAML.
+func NewInMemory(buf io.Reader) ParserProvider {
+	return &inMemoryProvider{buf: buf}
+}
 
-// Validate checks if the exporter configuration is valid
-func (cfg *Config) Validate() error {
-	return nil
+func (inp *inMemoryProvider) Get() (*config.Parser, error) {
+	v := config.NewViper()
+	v.SetConfigType("yaml")
+	if err := v.ReadConfig(inp.buf); err != nil {
+		return nil, err
+	}
+	return config.ParserFromViper(v), nil
 }