+++ CHANGELOG.md
@@ -6,6 +6,11 @@
 
 - Remove `configparser.DecodeTypeAndName`, use `config.IDFromString` (#2869)
 
+## 💡 Enhancements 💡
+
+- Add `validatable` interface with `Validate()` to all `config.<component>` (#2898)
+  - add the empty `Validate()` implementation for all component configs
+
 ## v0.24.0 Beta
 
 ## 🛑 Breaking changes 🛑